+++ arches_project/stylesheets/arches_styling.qss
