+++ arches_project/ui/edit_resource_add_confirmation_dialog.ui
