+++ arches_project/ui/edit_resource_replace_confirmation_dialog.ui
